test(data): 移动标题规范化测试
